09-Apr-2012: Also checks for Drupal install tar file. Downloads libraries. Downloads modules.

+++ __start.php
@@ -19,6 +19,37 @@ require_once '__start/archiver.inc';
 require_once '__start/system.archiver.inc';
 require_once '__start/system.tar.inc';
 
+// Processing drupal
+//$drupalPath = step1processDrupal();
+
+// Processing libraries
+$drupalPath = 'drupal-7.12/';
+//step2processLibraries($drupalPath, 'http://download.cksource.com/CKEditor/CKEditor/CKEditor%203.6.2/ckeditor_3.6.2.zip', 'ckeditor_3.6.2.zip');
+//step2processLibraries($drupalPath, 'http://css3pie.com/download-latest', 'PIE-1.0beta5.zip', 'PIE');
+
+// Processing Themes
+//step2processThemes($drupalPath, 'http://ftp.drupal.org/files/projects/omega-7.x-3.1.zip', 'omega-7.x-3.1.zip');
+
+// Processing Modules
+$modules = array(
+					//'http://ftp.drupal.org/files/projects/ckeditor-7.x-1.8.zip', 
+					//'http://ftp.drupal.org/files/projects/css3pie-7.x-2.1.zip', 
+					//'http://ftp.drupal.org/files/projects/webform-7.x-3.17.zip',  
+					//'http://ftp.drupal.org/files/projects/views-7.x-3.3.zip', 
+					'http://ftp.drupal.org/files/projects/ctools-7.x-1.0.zip',
+					'http://ftp.drupal.org/files/projects/admin_menu-7.x-3.0-rc1.zip',
+					'http://ftp.drupal.org/files/projects/token-7.x-1.0.zip',
+					'http://ftp.drupal.org/files/projects/pathauto-7.x-1.0.zip',
+				);
+
+foreach ($modules as $module) {
+	$filename = explode('/', $module);
+	step2processModules($drupalPath, $module, $filename[count($filename) - 1]);
+}
+
+////////////////////////
+// FUNCTIONS
+
 function drupal_mkdir($uri, $mode = NULL, $recursive = FALSE, $context = NULL) {
   if (!isset($mode)) {
     $mode = variable_get('file_chmod_directory', 0775);
@@ -45,14 +76,28 @@ function drupal_unlink($uri, $context = NULL) {
   }
 }
 
-$archZipFilename = 'drupal-7.12.tar.gz';
-$archTarFilename = 'drupal-7.12.tar';
-$drupalPath = '';
-
-if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org/files/projects/drupal-7.12.tar.gz')) !== false) {
-	print date('Y-m-d H:i:s') . "<br />";
-	print "Drupal 7.12 downloaded<br />";
+function step1processDrupal() {
+	$archZipFilename = 'drupal-7.12.tar.gz';
+	$archTarFilename = 'drupal-7.12.tar';
+	$drupalPath = '';
 	
+	if (!file_exists($archZipFilename)) {
+		// Download file from internet. Might not always work as require more than 30 sec
+		if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org/files/projects/drupal-7.12.tar.gz')) !== false) {
+			print date('Y-m-d H:i:s') . "<br />";
+			print "Drupal 7.12 downloaded<br />";
+		}
+		else {
+			print "Error: failed to download Drupal<br />";
+			exit;
+		}
+	}
+	else {
+		print "Notice: ".$archZipFilename." detected<br />";
+	}
+	
+	// Uncompressing
+		
 	file_put_contents($archTarFilename, file_get_contents("compress.zlib://" . $archZipFilename)); // get tar
 	//file_put_contents('/', file_get_contents("compress.zlib://drupal-7.12.tar"));
 	//system('tar -xvwzf drupal-7.12.tar.gz');
@@ -67,7 +112,7 @@ if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org
 	if (!$archiver) {
 		print (t('Cannot extract %file, not a valid archive.', array ('%file' => $file)));
 	}
-
+	
 	// Remove the directory if it exists, otherwise it might contain a mixture of
 	// old files mixed with the new files (e.g. in cases where files were removed
 	// from a later release).
@@ -106,7 +151,7 @@ if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org
 		
 	}
 	*/
-
+	
 	$directory = '';
 	//$extract_location = $directory . '/' . $project;
 	$extract_location = $directory . '/';
@@ -116,7 +161,7 @@ if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org
 		file_unmanaged_delete_recursive($extract_location);
 	}
 	*/
-
+	
 	//$project = strtok($files[0], '/\\');
 	//$archiver->getArchive()->addModify($files, $directory, $project);
 	//$archiver->getArchive()->addModify($files, $directory, $project);
@@ -132,33 +177,69 @@ if (file_put_contents($archZipFilename, file_get_contents('http://ftp.drupal.org
 	else
 	{
 		"Drupal install exists...<br />";
-	}		
-	unlink($archZipFilename);
-	unlink($archTarFilename);
+	}	
 	
-}
-else {
-	print "Error: failed to download Drupal<br />";
-}
-
-$ckzip = 'http://download.cksource.com/CKEditor/CKEditor/CKEditor%203.6.2/ckeditor_3.6.2.zip';
-$ckzipFile = 'ckeditor_3.6.2.zip';
-if (file_put_contents($ckzipFile, file_get_contents($ckzip)) !== false) {
-	print date('Y-m-d H:i:s') . "<br />";
-	print "CKEditor 3.6.2 downloaded<br />";
-	
-	$archiver = new ArchiverZip($ckzipFile);
-	if (!$archiver) {
-		print (t('Cannot extract %file, not a valid archive.', array ('%file' => $file)));
+	if (file_exists(unlink($archZipFilename))) {	
+		unlink($archZipFilename);
 	}
+	if (file_exists(unlink($archTarFilename))) {
+		unlink($archTarFilename);
+	}
+	
+	return $drupalPath;
+}
+	
+//////////////////////
+// CK Editor download
+//////////////////////
 
-	$directory = $drupalPath . 'sites/all/libraries/';
-	mkdir($drupalPath . 'sites/all/libraries/');
-	$files = $archiver->listContents();
-	$archiver->extract($directory);
-	//$archiver->getArchive()->extractModify($directory, 'ckeditor/_samples');
+function step2processLibraries($drupalPath = 'drupal-7.12/',
+						$ckzip = 'http://download.cksource.com/CKEditor/CKEditor/CKEditor%203.6.2/ckeditor_3.6.2.zip', 
+						$ckzipFile = 'ckeditor_3.6.2.zip', $extractFolder = '') {
+							
+		step2process($drupalPath, $ckzip, $ckzipFile, $extractFolder, 'libraries/');
 	
-	unlink($ckzipFile);
+}
+
+function step2processModules($drupalPath = 'drupal-7.12/',
+						$ckzip = 'http://ftp.drupal.org/files/projects/ckeditor-7.x-1.8.zip', 
+						$ckzipFile = 'ckeditor-7.x-1.8.zip', $extractFolder = '') {
+							
+		step2process($drupalPath, $ckzip, $ckzipFile, $extractFolder, 'modules/');
 	
-	print "CKEditor extraxted...<br />";
+}
+
+function step2processThemes($drupalPath = 'drupal-7.12/',
+						$ckzip = 'http://ftp.drupal.org/files/projects/omega-7.x-3.1.zip', 
+						$ckzipFile = 'omega-7.x-3.1.zip', $extractFolder = '') {
+							
+		step2process($drupalPath, $ckzip, $ckzipFile, $extractFolder, 'themes/');
+	
+}
+
+function step2process($drupalPath, $ckzip, $ckzipFile, $extractFolder, $mainSitesFolder) {
+	if (file_put_contents($ckzipFile, file_get_contents($ckzip)) !== false) {
+		print date('Y-m-d H:i:s') . "<br />";
+		print "File [".$ckzipFile."] downloaded<br />";
+		
+		$archiver = new ArchiverZip($ckzipFile);
+		if (!$archiver) {
+			print (t('Cannot extract %file, not a valid archive.', array ('%file' => $file)));
+		}
+	
+		$directory = $drupalPath . 'sites/all/' . $mainSitesFolder;
+		if (!file_exists($directory)) {
+			mkdir($directory);
+		}
+		if (!file_exists($directory.$extractFolder)) {
+			mkdir($directory.$extractFolder);
+		}
+		$files = $archiver->listContents();
+		$archiver->extract($directory.$extractFolder);
+		//$archiver->getArchive()->extractModify($directory, 'ckeditor/_samples');
+		
+		unlink($ckzipFile);
+		
+		print "Library [".$ckzipFile."] extraxted...<br />";
+	}
 }
